slight re-org
