+++ compiler/src/lib.rs
@@ -20,11 +20,13 @@ mod crypto;
 mod define;
 pub mod errors;
 mod fabric;
+mod fabric_util;
 mod lex;
 mod parser;
 mod policybuilder;
 mod protocols;
 mod ptypes;
 mod putil;
+mod weaver;
 mod zpl;
 mod zplstr;
